+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.1 (unreleased)
 
-- Added `var`, `std`, and `take` to vectors
+- Added `var`, `std`, `take`, `count`, and `length` to vectors
 - Added `sample` to data frames
 
 ## 0.1.0 (2020-05-13)