+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 ## 0.1.1 (unreleased)
 
-- Added `group`
-- Added `var`, `std`, `take`, `count`, and `length` to vectors
-- Added `sample` to data frames
+- Added `group` and `sample` to data frames
+- Added `tally`, `var`, `std`, `take`, `count`, and `length` to vectors
 - Improved error message for `read_csv` with no headers
 
 ## 0.1.0 (2020-05-13)