+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.1.1 (unreleased)
 
 - Added `group`, `one_hot` and `sample` to data frames
-- Added `tally`, `one_hot`, `var`, `std`, `take`, `count`, and `length` to vectors
+- Added `to`, `tally`, `one_hot`, `var`, `std`, `take`, `count`, and `length` to vectors
 - Improved error message for `read_csv` with no headers
 
 ## 0.1.0 (2020-05-13)