+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-## 0.1.1 (unreleased)
+## 0.1.1 (2020-06-10)
 
-- Added `group`, `one_hot` and `sample` to data frames
-- Added `to`, `tally`, `one_hot`, `var`, `std`, `take`, `count`, and `length` to vectors
+- Added methods and options for types
+- Added grouping
+- Added one-hot encoding
+- Added `sample` to data frames
+- Added `tally`, `var`, `std`, `take`, `count`, and `length` to vectors
 - Improved error message for `read_csv` with no headers
 
 ## 0.1.0 (2020-05-13)